updated logic and added validations in controller

+++ src/controllers/category.js
@@ -7,15 +7,13 @@ async function changeProductCategory(req, res) {
 
     if (!product_id) return res.status(400).json({ error: 'Missing product_id' });
 
-    console.log(req.query);
-
     try {
 
         await Category.findByIdAndUpdate(category_id, { $pull: { 'products': product_id } });
 
         const category = await Category.findByIdAndUpdate(category_id, { $push: { 'products': product_id } })
 
-        if (category == null) return res.status(400).json({ error: "Category Not founded" });
+        if (category == undefined) return res.status(400).json({ error: "Category Not founded" });
 
         await Product.findByIdAndUpdate(product_id, { category: category_id });
 
@@ -23,7 +21,7 @@ async function changeProductCategory(req, res) {
         res.json({ category });
 
     } catch (error) {
-        res.status(500).json({ error: 'Invalid category_id or product_id' });
+        res.status(400).json({ error: 'Invalid category_id or product_id' });
     }
 
 }
@@ -35,19 +33,17 @@ async function removeProductInCategory(req, res) {
 
     if (!product_id) return res.status(400).json({ error: 'Missing product_id' });
 
-    console.log(req.query);
-
     try {
         const category = await Category.findByIdAndUpdate(category_id, { $pull: { 'products': product_id } })
 
-        if (category == null) return res.status(400).json({ error: "Category Not founded" });
+        if (category == undefined) return res.status(400).json({ error: "Category Not founded" });
 
+        //Provisional empty category
         Product.findByIdAndUpdate(product_id, { category: "000000000000000000000000" });
-
 
         res.json({ category });
     } catch (error) {
-        res.status(500).json({ error: 'Invalid category_id or product_id' });
+        res.status(400).json({ error: 'Invalid category_id or product_id' });
     }
 
 
@@ -59,23 +55,29 @@ async function getCategory(req, res) {
 
     try {
         const category = await Category.findById(category_id);
+        if (category == undefined) return res.status(400).json({ error: 'category not founded' });
+
         res.json({ category });
     } catch (error) {
-        res.status(500).json({ error: 'Invalid category_id' });
+        res.status(400).json({ error: 'Invalid category_id' });
     }
 
 }
 
 async function createCategory(req, res) {
     const { name, } = req.body;
-    console.log(req.body);
     if (!name) return res.status(400).json({ error: 'Missing category data' });
 
     try {
+
+        const byName = await Category.findOne({ name });
+        if (byName) return res.status(400).json({ error: 'this category already exists' });
+
         const category = await Category.create({ name });
+
         res.json({ category });
     } catch (error) {
-        res.status(500).json({ error: 'Invalid category data' });
+        res.status(400).json({ error: 'Invalid category data' });
     }
 }
 
@@ -87,33 +89,34 @@ async function updateCategory(req, res) {
 
     try {
         const category = await Category.findById(category_id);
+        if (category == undefined) return res.status(500).json({ error: 'category not founded' });
+
         const data = {
             name: name || category.name,
         }
-        const updatedCategory = await User.findByIdAndUpdate(category_id, data, { new: true });
+        const updatedCategory = await Category.findByIdAndUpdate(category_id, data, { new: true });
         res.json({ product: updatedCategory });
     } catch (error) {
-        res.status(500).json({ error: 'Invalid category_id or data' });
+        res.status(400).json({ error: 'Invalid category_id or data' });
     }
 }
 
 async function deleteCategory(req, res) {
     const { category_id } = req.params;
-    console.log(req.params);
     if (!category_id) return res.status(400).json({ error: 'Missing category_id' });
 
 
     try {
         const category = await Category.findById(category_id);
+        if (category == undefined) return res.status(500).json({ error: 'category not founded' });
 
-        category.products.map(async product => await Product.findByIdAndUpdate({ "_id": product }, { category: "000000000000000000000000" }));
+        await Promise.all(category.products.map(async product => await Product.findByIdAndUpdate({ "_id": product }, { category: "000000000000000000000000" })));
 
         await Category.findByIdAndDelete(category_id);
 
         res.json({ category });
     } catch (error) {
-        console.log(error);
-        res.status(500).json({ error: 'Invalid category_id' });
+        res.status(400).json({ error: 'Invalid category_id' });
     }
 }
